+++ classes/NewsPlus.php
@@ -4,157 +4,95 @@
  * Contao Open Source CMS
  *
  * Copyright (c) 2015 Heimrich & Hannot GmbH
+ *
  * @package news_plus
- * @author Mathias Arzberger <develop@pdir.de>
+ * @author  Mathias Arzberger <develop@pdir.de>
  * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
  */
 
 namespace HeimrichHannot\NewsPlus;
 
-abstract class NewsPlus extends ModuleNewsPlus
+use Contao\News;
+use NewsCategories\NewsCategoryModel;
+
+class NewsPlus extends News
 {
-
-    /**
-     * get all news items by page pid
-     * @param array
-     * @param integer
-     * @param boolean
-     * @return array
-     */
-    public static function getAllNews($arrPages, $intRoot=0, $blnIsSitemap=false)
-    {
-        $arrRoot = array();
-
-        if ($intRoot > 0)
-        {
-            $arrRoot = \Database::getInstance()->getChildRecords($intRoot, 'tl_page');
-        }
-
-        $time = time();
-        $arrProcessed = array();
-
-        // Get all news archives
-        $objArchive = \NewsArchiveModel::findByProtected('');
-
-        // Walk through each archive
-        if ($objArchive !== null)
-        {
-            while ($objArchive->next())
-            {
-                // Skip news archives without target page
-                if (!$objArchive->jumpTo)
-                {
-                    continue;
-                }
-
-                // Skip news archives outside the root nodes
-                if (!empty($arrRoot) && !in_array($objArchive->jumpTo, $arrRoot))
-                {
-                    continue;
-                }
-
-                // Get the URL of the jumpTo page
-                if (!isset($arrProcessed[$objArchive->jumpTo]))
-                {
-                    $objParent = \PageModel::findWithDetails($objArchive->jumpTo);
-
-                    // The target page does not exist
-                    if ($objParent === null)
-                    {
-                        continue;
-                    }
-
-                    // The target page has not been published (see #5520)
-                    if (!$objParent->published || ($objParent->start != '' && $objParent->start > $time) || ($objParent->stop != '' && $objParent->stop < $time))
-                    {
-                        continue;
-                    }
-
-                    // The target page is exempt from the sitemap (see #6418)
-                    if ($blnIsSitemap && $objParent->sitemap == 'map_never')
-                    {
-                        continue;
-                    }
-
-                    // Set the domain (see #6421)
-                    // $domain = ($objParent->rootUseSSL ? 'https://' : 'http://') . ($objParent->domain ?: \Environment::get('host')) . TL_PATH . '/';
-
-                    // Generate the URL
-                    // $arrProcessed[$objArchive->jumpTo] = $domain . $this->generateFrontendUrl($objParent->row(), ((\Config::get('useAutoItem') && !\Config::get('disableAlias')) ?  '/%s' : '/items/%s'), $objParent->language);
-                }
-
-                $strUrl = $arrProcessed[$objArchive->jumpTo];
-
-                // Get the items
-                $objArticle = \NewsModel::findPublishedDefaultByPid($objArchive->id);
-
-                if ($objArticle !== null)
-                {
-                    while ($objArticle->next())
-                    {
-                        $arrPages[] = $objArticle->id;
-                    }
-                }
-            }
-        }
-
-        return $arrPages;
-    }
-
-
     public static function getAllPublishedNews($archives, $arrCategories)
     {
         $objAllArticles = NewsPlusModel::findPublishedByPidsAndCategories($archives, $arrCategories);
-        foreach($objAllArticles as $article){
+        foreach ($objAllArticles as $article)
+        {
             $arrIds[] = $article->id;
         }
+
         return $arrIds;
     }
 
-
     /**
-     * Calculate the span between two timestamps in days
-     * @param integer
-     * @param integer
-     * @return integer
+     * Adds support for overriding news archive's jump to in the primary news category assigned to the concrete news
+     *
+     * @param \NewsModel $objItem
+     * @param string     $strUrl
+     * @param string     $strBase
+     *
+     * @return string
      */
-    public static function calculateSpan($intStart, $intEnd)
+    protected function getLink($objItem, $strUrl, $strBase = '')
     {
-        return self::unixToJd($intEnd) - self::unixToJd($intStart);
-    }
-
-
-
-    /**
-     * Convert a UNIX timestamp to a Julian day
-     * @param integer
-     * @return integer
-     */
-    public static function unixToJd($tstamp)
-    {
-        list($year, $month, $day) = explode(',', date('Y,m,d', $tstamp));
-
-        // Make year a positive number
-        $year += ($year < 0 ? 4801 : 4800);
-
-        // Adjust the start of the year
-        if ($month > 2)
+        switch ($objItem->source)
         {
-            $month -= 3;
+            // Link to an external page
+            case 'external':
+                return $objItem->url;
+                break;
+
+            // Link to an internal page
+            case 'internal':
+                if (($objTarget = $objItem->getRelated('jumpTo')) !== null)
+                {
+                    /** @var \PageModel $objTarget */
+                    return $objTarget->getAbsoluteUrl();
+                }
+                break;
+
+            // Link to an article
+            case 'article':
+                if (($objArticle = \ArticleModel::findByPk($objItem->articleId, ['eager' => true])) !== null
+                    && ($objPid = $objArticle->getRelated('pid')) !== null
+                )
+                {
+                    /** @var \PageModel $objPid */
+                    return ampersand(
+                        $objPid->getAbsoluteUrl(
+                            '/articles/' . ((!\Config::get('disableAlias') && $objArticle->alias != '') ? $objArticle->alias : $objArticle->id)
+                        )
+                    );
+                }
+                break;
+
+            default:
+                $intJumpTo = 0;
+
+                // news category jump to override?
+                if ($objItem->primaryCategory && ($objCategory = NewsCategoryModel::findPublishedByIdOrAlias($objItem->primaryCategory)) !== null)
+                {
+                    $intJumpTo = $objCategory->jumpToDetails ?: $intJumpTo;
+                }
+
+                if (($objPage = \PageModel::findByPk($intJumpTo)) !== null)
+                {
+                    return $objPage->getAbsoluteUrl(((\Config::get('useAutoItem') && !\Config::get('disableAlias')) ?  '/' : '/items/') . ((!\Config::get('disableAlias') && $objItem->alias != '') ? $objItem->alias : $objItem->id));
+                }
+                break;
         }
-        else
+
+        // Backwards compatibility (see #8329)
+        if ($strBase != '' && !preg_match('#^https?://#', $strUrl))
         {
-            $month += 9;
-            --$year;
+            $strUrl = $strBase . $strUrl;
         }
 
-        $sdn  = floor((floor($year / 100) * 146097) / 4);
-        $sdn += floor((($year % 100) * 1461) / 4);
-        $sdn += floor(($month * 153 + 2) / 5);
-        $sdn += $day - 32045;
-
-        return $sdn;
+        // Link to the default page
+        return sprintf($strUrl, (($objItem->alias != '' && !\Config::get('disableAlias')) ? $objItem->alias : $objItem->id));
     }
-
-
 }